Following the In-class example:
AccData file is simply a file that stores the 3 x,y,z arrays and an
index

WriteFile is going to be the thread that actually writes to file [not
implemented]

I commented out all the step algorithm from StepDetector. Also added in
some code about how we need to add x,y,z values into the AccData and
objects BlockingQueue. [Note: it is commented out right now because it
isn't functional]